fix(doctor): coordinate migrations with the managed gateway
