+++ src/agents/worktrees/registry.test.ts
@@ -18,6 +18,8 @@ import {
   insertRegistryWorktreeProvisionedChunk,
   insertRegistryWorktree,
   listRegistryWorktrees,
+  listRegistryWorktreesForMigration,
+  hasLegacyRegistryWorktrees,
   updateRegistryWorktree,
 } from "./registry.js";
 import type { ManagedWorktreeRecord } from "./types.js";
@@ -35,6 +37,12 @@ describe("managed worktree registry", () => {
   afterEach(async () => {
     closeOpenClawStateDatabaseForTest();
     await fs.rm(root, { recursive: true, force: true });
+  });
+
+  it("inspects absent legacy worktrees without creating the state database", async () => {
+    expect(hasLegacyRegistryWorktrees(env)).toBe(false);
+    expect(listRegistryWorktreesForMigration(env)).toEqual([]);
+    await expect(fs.stat(env.OPENCLAW_STATE_DIR!)).rejects.toMatchObject({ code: "ENOENT" });
   });
 
   it("persists, orders, updates, and deletes worktree rows through Kysely", () => {
@@ -61,7 +69,9 @@ describe("managed worktree registry", () => {
       lastActiveAt: 20,
     });
 
+    expect(hasLegacyRegistryWorktrees(env)).toBe(true);
     expect(listRegistryWorktrees(env).map((entry) => entry.id)).toEqual(["second", "first"]);
+    expect(listRegistryWorktreesForMigration(env)).toEqual(listRegistryWorktrees(env));
     expect(findLiveRegistryWorktreeByPath(env, record.path)).toMatchObject({
       id: "first",
       ownerKind: "workboard",
