+++ src/flows/doctor-health.ts
@@ -89,84 +89,107 @@ export async function runDoctorHealthFlow(runtime?: RuntimeEnv, options: DoctorO
     assertConfigWriteAllowedInCurrentMode();
   }
 
-  // Keep side-effect-heavy legacy checks before structured contributions until fully migrated.
-  const { maybeRepairUiProtocolFreshness } = await import("../commands/doctor-ui.js");
-  const { noteSourceInstallIssues } = await import("../commands/doctor-install.js");
-  const { noteStalePluginRuntimeSymlinks } =
-    await import("../commands/doctor/shared/plugin-runtime-symlinks.js");
-  const { noteStartupOptimizationHints } = await import("../commands/doctor-platform-notes.js");
-  await maybeRepairUiProtocolFreshness(effectiveRuntime, prompter);
-  noteSourceInstallIssues(root);
-  await noteStalePluginRuntimeSymlinks(root);
-  noteStartupOptimizationHints();
+  const { beginDoctorMaintenance } = await import("../commands/doctor-maintenance.js");
+  const maintenance = await beginDoctorMaintenance({ options, root, runtime: effectiveRuntime });
+  let exitCode: number | undefined;
+  try {
+    // Keep side-effect-heavy legacy checks before structured contributions until fully migrated.
+    const { maybeRepairUiProtocolFreshness } = await import("../commands/doctor-ui.js");
+    const { noteSourceInstallIssues } = await import("../commands/doctor-install.js");
+    const { noteStalePluginRuntimeSymlinks } =
+      await import("../commands/doctor/shared/plugin-runtime-symlinks.js");
+    const { noteStartupOptimizationHints } = await import("../commands/doctor-platform-notes.js");
+    await maybeRepairUiProtocolFreshness(effectiveRuntime, prompter);
+    noteSourceInstallIssues(root);
+    await noteStalePluginRuntimeSymlinks(root);
+    noteStartupOptimizationHints();
 
-  const { loadAndMaybeMigrateDoctorConfig } = await import("../commands/doctor-config-flow.js");
-  const configResult = await loadAndMaybeMigrateDoctorConfig({
-    options,
-    confirm: (p) => prompter.confirm(p),
-    runtime: effectiveRuntime,
-    prompter,
-  });
-  const { CONFIG_PATH } = await loadConfigModule();
-  const ctx: DoctorHealthFlowContext = {
-    runtime: effectiveRuntime,
-    options,
-    prompter,
-    configResult,
-    cfg: configResult.cfg,
-    cfgForPersistence: structuredClone(configResult.cfg),
-    sourceConfigValid: configResult.sourceConfigValid ?? true,
-    configPath: configResult.path ?? CONFIG_PATH,
-    stateDirExistedAtStart,
-    runWithPluginMetadataSnapshot: configResult.runWithPluginMetadataSnapshot,
-    invalidatePluginMetadataSnapshot: configResult.invalidatePluginMetadataSnapshot,
-  };
-  const { runDoctorHealthContributions } = await import("./doctor-health-contributions.js");
-  await runDoctorHealthContributions(ctx);
-  if (ctx.configWriteRefusal) {
-    // Config fixes were computed but refused by the writer; the warning above
-    // already lists the manual work. This failure outranks a recoverable
-    // post-install advisory because the run did not converge.
-    outro(
-      ctx.configResultWriteCommitted === true
-        ? "Doctor finished, but some config fixes were not applied."
-        : "Doctor finished, but config fixes were not applied.",
-    );
-    effectiveRuntime.exit(1);
-    return;
-  }
-  if (options.repair === true || options.yes === true) {
-    // Migration warnings also cover optional archives; certify required runtime
-    // schemas independently before reporting success or a recoverable advisory.
-    const { assertOpenClawDatabasesReady } =
-      await import("../state/openclaw-database-preflight.js");
-    const { resolveConfiguredAgentDatabaseTargets } = await import("../config/sessions/targets.js");
-    assertOpenClawDatabasesReady({
-      env: process.env,
-      operation: "doctor",
-      configuredAgentDatabaseTargets: resolveConfiguredAgentDatabaseTargets(ctx.cfg, {
-        env: process.env,
-      }),
+    const { loadAndMaybeMigrateDoctorConfig } = await import("../commands/doctor-config-flow.js");
+    const configResult = await loadAndMaybeMigrateDoctorConfig({
+      options,
+      confirm: (p) => prompter.confirm(p),
+      runtime: effectiveRuntime,
+      prompter,
     });
-    const { assertConfiguredWorkspaceStateReady } =
-      await import("../agents/workspace-state-dirs.js");
-    assertConfiguredWorkspaceStateReady({ cfg: ctx.cfg });
-  }
-  if (ctx.postInstallDoctorResult) {
-    const {
-      UPDATE_POST_INSTALL_DOCTOR_ADVISORY_EXIT_CODE,
-      UPDATE_POST_INSTALL_DOCTOR_RESULT_PATH_ENV,
-      writeUpdatePostInstallDoctorResult,
-    } = await import("../infra/update-doctor-result.js");
-    const resultPath = process.env[UPDATE_POST_INSTALL_DOCTOR_RESULT_PATH_ENV]?.trim();
-    if (resultPath) {
-      await writeUpdatePostInstallDoctorResult({
-        resultPath,
-        result: ctx.postInstallDoctorResult,
-      });
-      effectiveRuntime.exit(UPDATE_POST_INSTALL_DOCTOR_ADVISORY_EXIT_CODE);
+    const { CONFIG_PATH } = await loadConfigModule();
+    const ctx: DoctorHealthFlowContext = {
+      runtime: effectiveRuntime,
+      options,
+      prompter,
+      configResult,
+      cfg: configResult.cfg,
+      cfgForPersistence: structuredClone(configResult.cfg),
+      sourceConfigValid: configResult.sourceConfigValid ?? true,
+      configPath: configResult.path ?? CONFIG_PATH,
+      stateDirExistedAtStart,
+      gatewayMaintenanceActive: maintenance !== undefined,
+      runWithPluginMetadataSnapshot: configResult.runWithPluginMetadataSnapshot,
+      invalidatePluginMetadataSnapshot: configResult.invalidatePluginMetadataSnapshot,
+    };
+    const { runDoctorHealthContributions } = await import("./doctor-health-contributions.js");
+    await runDoctorHealthContributions(ctx);
+    if (ctx.configWriteRefusal) {
+      // Config fixes were computed but refused by the writer; the warning above
+      // already lists the manual work. This failure outranks a recoverable
+      // post-install advisory because the run did not converge.
+      outro(
+        ctx.configResultWriteCommitted === true
+          ? "Doctor finished, but some config fixes were not applied."
+          : "Doctor finished, but config fixes were not applied.",
+      );
+      exitCode = 1;
       return;
     }
+    if (options.repair === true || options.yes === true) {
+      // Migration warnings also cover optional archives; certify required runtime
+      // schemas independently before reporting success or a recoverable advisory.
+      const { assertOpenClawDatabasesReady } =
+        await import("../state/openclaw-database-preflight.js");
+      const { resolveConfiguredAgentDatabaseTargets } =
+        await import("../config/sessions/targets.js");
+      assertOpenClawDatabasesReady({
+        env: process.env,
+        operation: "doctor",
+        configuredAgentDatabaseTargets: resolveConfiguredAgentDatabaseTargets(ctx.cfg, {
+          env: process.env,
+        }),
+      });
+      const { assertConfiguredWorkspaceStateReady } =
+        await import("../agents/workspace-state-dirs.js");
+      assertConfiguredWorkspaceStateReady({ cfg: ctx.cfg });
+    }
+    await maintenance?.finish(ctx.cfg);
+    if (ctx.postInstallDoctorResult) {
+      const {
+        UPDATE_POST_INSTALL_DOCTOR_ADVISORY_EXIT_CODE,
+        UPDATE_POST_INSTALL_DOCTOR_RESULT_PATH_ENV,
+        writeUpdatePostInstallDoctorResult,
+      } = await import("../infra/update-doctor-result.js");
+      const resultPath = process.env[UPDATE_POST_INSTALL_DOCTOR_RESULT_PATH_ENV]?.trim();
+      if (resultPath) {
+        await writeUpdatePostInstallDoctorResult({
+          resultPath,
+          result: ctx.postInstallDoctorResult,
+        });
+        exitCode = UPDATE_POST_INSTALL_DOCTOR_ADVISORY_EXIT_CODE;
+        return;
+      }
+    }
+  } catch (error) {
+    if (maintenance) {
+      effectiveRuntime.error(
+        "Doctor could not complete maintenance. Check the reported service state, resolve the failure, and rerun doctor --fix.",
+      );
+    }
+    throw error;
+  } finally {
+    await maintenance?.release();
+    // The default runtime exits synchronously; finish native recovery and release
+    // maintenance leases before handing it an exit code.
+    if (exitCode !== undefined) {
+      effectiveRuntime.exit(exitCode);
+    }
   }
+
   outro("Doctor complete.");
 }
